+++ autogpt_platform/backend/backend/data/user.py
@@ -1,15 +1,11 @@
-import json
-from cryptography.fernet import Fernet
-
 from typing import Optional
 
 from autogpt_libs.supabase_integration_credentials_store.types import UserMetadataRaw
-from backend.util.settings import Settings
 from fastapi import HTTPException
-from prisma import Json
 from prisma.models import User
 
 from backend.data.db import prisma
+from backend.util.encryption import EncryptedJson
 
 DEFAULT_USER_ID = "3e53486c-cf57-477e-ba2a-cb02dc828e1a"
 DEFAULT_EMAIL = "default@example.com"
@@ -54,45 +50,16 @@ async def create_default_user() -> Optional[User]:
     return User.model_validate(user)
 
 
-ENCRYPTION_KEY = Settings().secrets.encryption_key
-
-class EncryptedJson:
-    def __init__(self, key: Optional[str] = None):
-        # Use provided key or get from environment
-        self.key = key or ENCRYPTION_KEY
-        if not self.key:
-            raise ValueError(
-                "Encryption key must be provided or set in ENCRYPTION_KEY environment variable"
-            )
-        self.fernet = Fernet(
-            self.key.encode() if isinstance(self.key, str) else self.key
-        )
-
-    def encrypt(self, data: dict) -> str:
-        """Encrypt dictionary data to string"""
-        json_str = json.dumps(data)
-        encrypted = self.fernet.encrypt(json_str.encode())
-        return encrypted.decode()
-
-    def decrypt(self, encrypted_str: str) -> dict:
-        """Decrypt string to dictionary"""
-        if not encrypted_str:
-            return {}
-        decrypted = self.fernet.decrypt(encrypted_str.encode())
-        return json.loads(decrypted.decode())
-
-
-# Initialize encryption
-json_encryptor = EncryptedJson()
-
-
 def encrypt_metadata(metadata: UserMetadataRaw) -> str:
     """Encrypt metadata Pydantic model to string"""
+    # Initialize encryption
+    json_encryptor = EncryptedJson()
     return json_encryptor.encrypt(metadata.model_dump())
 
 
 def decrypt_metadata(metadata: str) -> UserMetadataRaw:
     """Decrypt string to metadata Pydantic model"""
+    json_encryptor = EncryptedJson()
     decrypted_dict = json_encryptor.decrypt(metadata)
     return UserMetadataRaw.model_validate(decrypted_dict)
 
@@ -110,6 +77,7 @@ async def get_user_metadata(user_id: str) -> UserMetadataRaw:
     else:
         decrypted_metadata = decrypt_metadata(metadata)
         return decrypted_metadata
+
 
 async def update_user_metadata(user_id: str, metadata: UserMetadataRaw):
     # Encrypt the metadata
